Rename module and remove debugging tic-toc timer
